+++ action_server/tests/action_server_tests/resources/no_conda/check_headers/headers_action.py
@@ -1,24 +1,19 @@
-# import os
-# from pathlib import Path
-#
-# from robocorp.actions import Request, action
-#
-#
-# @action
-# def check_headers(request: Request, name: str, title="Mr.") -> str:
-#     """
-#     Gets something from the headers.
-#
-#     Args:
-#         name: The name of the person to greet.
-#         title: The title for the persor (Mr., Mrs., ...).
-#
-#     Returns:
-#         The greeting for the person.
-#     """
-#     artifacts = Path(os.environ["ROBOT_ARTIFACTS"])
-#     subdir = artifacts / "subdir"
-#     subdir.mkdir(parents=True, exist_ok=True)
-#     txt = subdir / "myfile.txt"
-#     txt.write_text("Some text in myfile.txt")
-#     return f"Hello {title} {name}."
+import json
+
+from robocorp.actions import Request, action
+
+
+@action
+def check_headers(request: Request, name: str, title="Mr.") -> str:
+    """
+    Gets something from the headers.
+
+    Args:
+        name: The name of the person to greet.
+        title: The title for the persor (Mr., Mrs., ...).
+
+    Returns:
+        The greeting for the person.
+    """
+    ret = {"headers": dict(request.headers), "cookies": dict(request.cookies)}
+    return json.dumps(ret)