+++ actions/src/robocorp/actions/_protocols.py
@@ -1,9 +1,14 @@
+import typing
 from typing import Any, Callable, Sequence
 
 from robocorp.tasks import ITask as _ITask
 from robocorp.tasks import Status as _Status
 
-IAction = _ITask
+
+class IAction(_ITask, typing.Protocol):
+    pass
+
+
 Status = _Status
 
 IActionCallback = Callable[[IAction], Any]